+++ R/plot.parameters_pca.R
@@ -1,4 +1,3 @@
-#' @importFrom dplyr group_by mutate ungroup select one_of n
 #' @export
 data_plot.parameters_pca <- function(x, data = NULL, ...){
   dataplot <- as.data.frame(x)
@@ -58,9 +57,9 @@ plot.see_parameters_pca <- function(x, type = c("bar", "line"), text_size = 3.5,
     )
   }
 
-  p <- x %>%
-    as.data.frame() %>%
-    ggplot(aes(
+  p <- ggplot(
+    as.data.frame(x),
+    aes(
       x = .data$Variable,
       y = abs(.data$y),
       fill = .data$y,