+++ R/plot.rope.R
@@ -1,4 +1,3 @@
-#' @importFrom dplyr group_by mutate ungroup select one_of n
 #' @export
 data_plot.rope <- function(x, data = NULL, show_intercept = FALSE, ...){
   if (is.null(data)) {
@@ -86,9 +85,9 @@ plot.see_rope <- function(x, data = NULL, rope_alpha = 0.5, rope_color = "cadetb
   # get labels
   labels <- .clean_parameter_names(x$y, grid = !is.null(n_columns))
 
-  p <- x %>%
-    as.data.frame() %>%
-    ggplot(aes(
+  p <- ggplot(
+    as.data.frame(x),
+    aes(
       x = .data$x,
       y = .data$y,
       height = .data$height,