+++ NEWS.md
@@ -1,3 +1,20 @@
+# see 0.4.1
+
+## General
+
+* Reduce package dependencies.
+
+## New plot-functions
+
+* Added plot-function for `correlation::correlation()`.
+
+## Changes to functions
+
+* `plot()` for `model_parameters.rma()` now arranges facets by subgroups (see vignettes).
+* minor improvements to `plot()` for `parameters::cluster_analysis()`.
+* minor improvements to `plot()` for `parameters::model_parameters.brmsfit()`.
+* Plot-function for `bayestestR::si()` gets a `support_only` argument, to lot only the support data or the "raw" prior and posterior distributions.
+
 # see 0.4.0
 
 ## Breaking changes